[NativeFS] Refactor WPT tests to support native FS as well.
Also changes the sandboxed file system version of these tests to run in
both windows and workers.

The native file system version of these tests are currently manual tests.
In the future we can automate them. Also currently some of the writer
tests don't pass in the native file system version, because of things like
bug 998913.

Bug: 998917
Change-Id: I4aefee6e35008afd73c2cf316d16ea189717bd4b