[LayoutNG] Fix incorrect rounding in ShapingLineBreaker
This patch fixes the rounding the end position to the same
direction as the rounding the start position. Before this
change, the end position was rounded to the line-end
direction. This may have resulted in slightly larger
available space than the given one.

Such an error can cause a large visual differences, because
|NGLineBreaker| determines the overflow by comparing the
result width to the available width, and if overflow occurs,
it assumes that it cannot break the result further.

This patch also adds |ShapingLineBreaker::Result::is_overflow|
so that |ShapingLineBreaker| can communicate its overflow
status to the caller. This ensures to determine the overflow
status correctly even if similar error occurs.

Bug: 1002442
Change-Id: Iebee80117c9930eade029c45c561236b8916981a