Fix a bug that could allow duplicate form submissions
Previous to this CL, and after [1], if a form Submit button had an
onclick handler that also called form.submit() and did not call
event.preventDefault(), the form would get submitted twice. The second
request was eventually cancelled, but not before it hit the network.
Note that this behavior is not specified - see [2]. Prior browser
behavior (including Firefox, Safari, and Chrome pre-[1]) was to only
submit the form once. So this CL returns to that behavior: only a
single submission will be made.

[1] https://chromium.googlesource.com/chromium/src/+/6931ab86f19aa79abbdd0c1062084e16b5c4f0f6
[2] https://www.w3.org/TR/html52/sec-forms.html#forms-form-submission

Bug: 977882
Change-Id: Iaef0067fa9136c074ab94b1c7baa95cff5c145ce